Improve watch mode development section and fix incorrect link
The package `serve` is incorrectly linked to the repository tj/serve.
This is because @tj transferred ownership of the `serve` package on the
npm registry to zeit/serve. Thus, the link and the installation command
correspond to two different packages.

The link now points to zeit/serve, which has similar functionality to
tj/serve, in addition to an option for effectively serving single page
apps.

Additionally, the instructions have been improved to instruct readers to
install the package locally and run it using an npm script as opposed to
installing it globally.