Clarify how to include CSS modules
This clarification is necessary because users might try to copy and paste the below code, and will run into bugs that do not have an obvious solution. If users set `modules: true` without meaning to, they will run into an issue where the classes will be replaced with generated ones, but this isn't made clear by the documentation. 

This PR also fixes some grammar mistakes for clarity.