Change externals module contexts from `global` to root
The currently deployed documentation indicates that there are 4 module contexts you can specify under the externals config - `global`, `commonjs`, `commonjs2`, and `amd`, which contradicts both the examples that surround it, and the actual library functionality. 

This just changes the word "global" to "root" in the section that lists the module contexts.