docs(guides): rework the caching guide (#1436)
Rewrite this guide to synchronize it with the other guides as per
#1258. Also include some content and scenarios based on various
articles written on the subject. Note that we will likely follow up 
with a few more PRs to elaborate on this base.

Markdown Issue

The issue is being thrown because of a recent change to the
uglifyjs-webpack-plugin. We'll probably need to put a more stable
fix in place (if not that table will continue to be broken). It looks
like maybe the cause is multiple escaped `|` characters used within
a single cell.