docs(guides): add build-performance page (#1506)
This adds a solid baseline for how to start improving build performance
with specific pointers for development and production. It can be built on
in other PRs but probably doesn't need to be synchronized with the earlier
guides (as it's a lot of one off changes that depend on your use case).