chore: upgrade webpack/antwar/deps (#980)
The main purpose of this PR is just to bring our dependencies,
especially webpack and antwar up to date. In the process of
doing this, we also resolved a variety of bugs and other issues.
There is still much work to be done but this is definitely a step
in the right direction and should make further changes easier.
Here are some other notable changes:

* refactor(components): clean up formatting and usage
* fix(components): fix `PropTypes` warnings
* refactor(config): remove unused vote routes
* refactor(config): utilize `disable` option in the extract plugin to simplify config
* fix(build-process): fix missing `<title>` tags
* fix(components): simplify index page management in sidebar (#1085)
* docs: minor improvements to index pages
* fix(tests): ignore npm-install-weback-plugin

Ignore the "sucks" mention in the npm-install-webpack-plugin file
for now. It isn't used in a derogatory way so I'm not sure we should
change it but I'm also not sure we should `allow` "sucks" globally.