docs: general formatting clean up
Mostly clean up of tables so they're a little less noisy and easier
to read. I'm slowly trying to enforce a single syntax for tables
throughout the whole site (would be great if we could lint this).
Also formatted certain things as code and cleaned up wording
in a few places.