Linking compiler docs on plugins page
While authoring plugins the list of hooks is essential knowledge, so link to hooks documentation page should be accessible in the top of plugins authoring guide.
The exact location is not very important, but it should be easy to find.

(for newcomer like myself it was not easy to realize the link between compiler api and plugins at first)