Updated Usage Instructions with Fragment Links
This is just a quality of life adjustment that updates the list of ways to use loaders with fragment links to the relevant section in the docs.  In their current state, the section feels like a dead end, abruptly cutting off with three bullet points and no examples. While one can read on and figure it out, there is a break in focus that is quite distracting. No one like to have a "wtf" moment when reading docs so this hopefully reduces that friction and helps people connect the dots.