style(vote): add some very simple header styling for consistency
We may be able to figure out a cleaner way to do this but for now this keeps
things pretty consistent with the rest of the site. In general the voting app should
probably do most of it's own styling though -- however we should probably
include the same `reset` stylesheet in both places for consistency.