Add note on how plugins can intercept stats
I was trying to suppress some stats in my project which used the webpack Node API, so I set my stats object according to that documentation. It wasn't working, and I had no idea why. That's until I asked on SO and was told webpack-dev-middleware intercepts stats. It's not obvious that happens, so I want to add this note for devs to be aware of such things. 

SO answer: https://stackoverflow.com/a/48453063/3121196