docs(v4): webpack 4 documentation (#1856)
* adding docs for new `output.devtoolNamespace` option and [namespace] placeholder

* adding extra lines as indicated

* docs(config): update `hashFunction` docs based on v4 (#1761)

* docs(api): reformat index page at 80 characters / line

* docs(api): reformat node api docs and add more details on the compiler

This moves some of the content from `compiler.md` and brings it into the
main node api documentation in order to allow `compiler.md` to become
`compiler-hooks.md` which fits better in the context of plugins.

* docs(api): reformat/rewrite the lead-in plugin docs and remove `tapable.md`

This moves some of the `Tapable` discussion to the lead-in `plugins.md` to
prevent doc duplication. Instead of redocumenting the `tapable` package, we
should focus on improving the README and simply link to there throughout
our docs. As I've mentioned before, we can also create a "Utilities" section
for things other than "Loaders" and "Plugins" that we want to dynamically
pull in.

* docs(api): remove `module-factories.md` and `template.md`

These guides were full of todos and probably lead to more confusion
than clarity. We can discuss more and add them back in once there's
a clear flow for how to incorporate and document them.

* docs(api): reformat line length in plugins.md

* docs(api): rewrite, rename, and resort the `compiler.md` page

This page now focuses specifically on the `hooks` as all the non-plugin related
content was moved to `node.md`. I updated the hooks to the new syntax, added
all undocumented hooks, reformatted each hook as a section to allow more
breathing room, and clarified exactly which `tapable` hook is used for each
section. Still need to get more info on some of the new hooks before we can
ship this...

* docs(api): rewrite, rename, and resort the `compilation.md` page

This page now uses the same section setup as `compiler-hooks.md` and
is up to date with the latest hooks available in the webpack 4 alpha release.

* docs(api): fix sorting and start to reformat the last two plugin pages

* docs(api): begin rewrite of the resolver page

* docs(config): fix header nesting in resolve.md

* docs(api): rename resolvers page for clarity and finish rewrite

* docs(api): make minor corrections to plugin docs

* docs(api): rewrite the parser documentation

Improve lead-in to give a little more context on what the parser does. Add
the latest hooks and use the same layout as the other pages in the section.

* docs(api): update some of the compiler hook descriptions

* docs(plugins): add `ProfilingPlugin` documentation (#1830)

* docs(config): document `info-verbosity` option in watch.md (#1817)

* update tree shaking guide for webpack 4

* docs(Guides): Update Getting Started to use webpack 4

Updates the Get Started guide to use webpack 4 and adds some additional information about npm.

* docs(concepts): Update plugins.md (#1871)

Update the plugins page in the concepts section to be compatible with webpack 4. Also adds a note about referencing the plugin name in camel-case when calling the `tap` method.

* docs(Stats): add `buitlAt` to the example configuration (#1878)

This adds the `builtAt` option to the example configuration.

* docs(guides): add doctype to Get Started (#1874)

This adds valid doctypes to all the HTML examples in the Get Started guide.

* docs(API): add missing config options to cli.md (#1880)

This adds a table of config options for v4 to the API documentation.

* docs(Concepts): update concepts page for v4 (#1883)

This updates the Concepts page to facilitate webpack v4's options and changes.

* update related links to be relevant for webpack 4

* add note about side effects and non-JS files

* add some more advanced usage notes

* docs(concepts) Clarification about NODE_ENV and mode (#1885)

* chore(site): Delete package-lock.json (#1917)

* chore: clean package-lock.json

* chore: sync yarn.lock

* docs(guides): Correct a path error in getting started example (#1899)

* correct a path error in example

* chore: Resolve conflicts on next branch (#1923)

* Update gitignore

* Resolve conflicts

* Resolve conflicts

* docs(api) Fix markdown lint errors

* docs(api) Fix Plugins wording