Document "none" values for mode
The schema validator supports a "none" value for `mode`, but it is not documented.  It is helpful if you want to control your own optimization, but not get the warning message or the default optimization behavior when `mode` is not set.

I was also confused by the following:
```
T> Please remember that setting `NODE_ENV` doesn't automatically set `mode`.
```

When I remove "mode", but still set `NODE_ENV=production`, my bundle is getting minified.  Is this a bug, something wrong in my config, or is the above line unclear / incorrect?