chore(rebuild): move redirect generation to prod to improve performance
Including this plugin in `common` causes it to be run twice for production builds as both
parallelized configs pick this up. We only need it in `prod` and it's kind of nice to have
all html generation happen in one build so lumping it together with the `SSGPlugin` makes
sense.