Improve runtimeChunk documentation
I needed to use runtimeChunk in a way that was not documented,
so I looked at the webpack source code and figured out how to
use this option better. I figured it would be good to improve the
docs for future explorers.

This is the code I was looking at:

  https://github.com/webpack/webpack/blob/941be2941e5fa51136454158ea812000c78c4a08/lib/WebpackOptionsDefaulter.js#L234-L246