Update externals.md
Give an example of where multiple libraries can be defined. I've used react because it typically requires multiple libraries to be included and there are no examples of multiple libraries being defined as external. Externals is a plural after all so implies multiple libraries can be excluded from the build.