docs(guides) Making the examples consistent on Getting started (#2259)
* Making the examples consistent.

There was some inconsistency between referring to main.js in the html file, and bundle.js on other places. I assume that the default in previous versions was bundle.js, however following the examples with console output, webpack uses main.js, so I have updated the examples to reflect this, as it reflected quite badly that the getting started instructions followed exactly as examples presented wouldn't actually work. Happy to have my name removed as a contributed if deemed necessary, it's only a small change after all.

* Restoring/removing line breakes after missed before PR, also changed the wording of a sentence to remove brackets and make the intent clearer

* Removing yet more line breaks that were unintentionally added

* Correct a directory naming error, and removing redundant text