Fix Sidebar order alphabetically (#2288)
Pages sidebar order was incorrect in all pages, this PR fixes it.

We are using `directory-tree` to get a JS Object representation of our content, at this moment it returns `index.md` files as any other page, but we have to use it as our main anchor in the sidebar.

I tried to change how `sort` function works on  `content-tree-enhancers.js`, without success, as last resort I manually moved the item to the beginning of the array. It works but it is a hack.

We should think in a better structure for our content, or at least create a different representation between system files and site content structure and navigation (sidebar).