Add a warning note for common errors when specifying rules
When specifying a test in `module.rules`, webpack accepts either a regex or string. However it's easy to mistakenly qoute the regex which will cause webpack to not use the specified loader. I added this warning to help newcomers like me who might  stumbled into this situation save themselves frustrations and wasted time.