Update hot-module-replacement.md
Only added a parenthesis tip that will hopefully help future readers getting stuck on the same part I did. The tutorial updates the button.onclick printMe() only a couple of steps after enabling HMR and a focused user might think the code is not working when the output from the button click to the console.log does not update with the HMR changes.