Specify the command to set a custom output
Suggested a modification in the documentation with instructions on how to set an output file different from the default one, when running webpack from the command line. 
In the paragraph about creating a bundle, there was a description on how to run the webpack binary that would use the default entry point and output, but there wasn't additional information about a custom entry or output and the flag needed to specify the output file.