Small readme improvements, remove the word "simple"
Most of us are reading the docs because we don't know how to do something, and it's not fun to hear that the concept we're having a hard time understanding is "simple" or "easy". I'd love you to show me how simple it is instead. 

A great reference: https://css-tricks.com/words-avoid-educational-writing/

Updated some language around `contenthash` as well, hopefully it's a little more direct. Let me know what you think.