docs: added more info for magic comments
docs: added title of Magic Comments so it can be found in the search bar. Added preload and prefetch as although they are in the guide I feel it would be beneficial to have all the magic comments in the one place and a link to the guide for more. Any suggestions just let me know