Added CSS rules for printing; added script to create chapter-wide documentation files from single files.
Currently, it's not possible to create a decent printed (or PDF) version of the documentation:

(a) Printed pages show all the navigation and interactive chrome, thereby unnecessarily wasting space.
(b) Printing all of a chapter's pages to get a comprehensive lecture wastes a lot of space when
    each of a chapter's pages gets printed separately.

This commit fixes both issues:

(a) New CSS rules have been added to remove navigation and interactive chrome when printing a
    documentation page.
(b) A Node script has been added to concatenate a chapter's Markdown documents into a single page.

To do:

The webpack documentation team is suggested to add the new Node script to the build script and to
publish the generated concatenated content, converted to HTML, on their website for printing.