Update dev-server.md
This gave me a cryptic error, and I spent a long time trying to figure out why it was happening. I'm sure this would affect others as well.

When passing in the second argument to `webpack(config, callback)`, it will return a different object then when called with just the config. This causes problems when using webpack-dev-server because it's expecting a compiler, not the webpack config with a `compiler` prop.

You would see the error `TypeError: Cannot read property 'plugins' of undefined` when the callback is added to `webpack()`.