Update caching.md
There is an error in the punctuation of the following compound/complex sentence:
`This allows sites to load faster with less unnecessary network traffic. However, it can also cause headaches when you need new code to be picked up.` I propose the grammar changes in the `caching.md`. The following sentence would also be correct. 

`This allows sites to load faster with less unnecessary network traffic; however, it can also cause headaches when you need new code to be picked up.`