docs(plugins): Refer to optimization configuration
I ran across a case where I wanted to disable the default `concatenateModules: true` in production mode, read these docs and came to the conclusion there wasn't an available option. Maybe it's obvious in retrospect, but figured it might be helpful for someone else to find the reference in the  module-concatenation-plugin docs here.