Update lazy-loading.md
I had to add the CleanWebpackPlugin and the HtmlWebpackPlugin in the webpack.config.js file in order to get the same results that were shown in the example after the npm run build command was ran.  These two plugins were not present in the Code Splitting subsection of the Guides section.