docs(migrate): Update 5.md (#4474)
* Update 5.md

`--dev` is deprecated. npm 6 suggest to use `--also=dev` and npm 7 suggests to use "--include=dev". I suggest to simply remove that flag because if webpack is already in devDependencies, it will be updated in any case.

Even using `--save-dev` works but it will also move webpack from `dependencies` to `devDependencies`, and maybe it's not always what the developer wants.

* Update src/content/migrate/5.md

Co-authored-by: Sam Chen <chenxsan@gmail.com>

Co-authored-by: Sam Chen <chenxsan@gmail.com>