Set "moduleResolution" to "node"
By default, "moduleResolution" is set to "classic", which does not allow importing lodash (or any module) in the way described in this guide.
Setting it to "node" instead fixes the issue. This setting is the industry standard; https://www.typescriptlang.org/docs/handbook/module-resolution.html

P.S. webpack is awesome, and the docs are great, many thanks!