docs(concepts): Small improvements to readability in index.mdx (#5311)
* Small improvements to readability in index.mdx

Added a brief blurb to explain what bundles are. Rewording to indicate the relationship between the dependency graph and bundle generation.

The first warning under Loaders was confusing because it references an ability to import modules as if this was mentioned previously in the article. It wasn't. Rewording to make webpack the subject of the sentence.

Rewording to show that the html-webpack-plugin in the code example injects the bundles into the generated html file. Before, the article made it seem as if injecting bundles was somehow a necessary step to creating an html file.

* Update src/content/concepts/index.mdx

Co-authored-by: Ryan Greyling <57382894+RyanGreyling2@users.noreply.github.com>

Co-authored-by: Sam Chen <chenxsan@gmail.com>