Update docs for entry-context
I think this current directory has a ambiguity. Does the current directory refer to the root directory of the project? Or is it the directory where webpack.config.js is located?

After I tested it, I found neither of them, the default should be the directory of the script that starts webpack. For example, I ran the script `webpack --config relativebase` in a non-project root directory. Then, the value of the context at this time is the directory where the script is located.

In order to better illustrate my ideas, I wrote a demo code, but I don’t know how to give the sample code, so I created a warehouse to store the sample code. https://github.com/spicylemonhaha/webpack-demo