Update docs for entry-context
First of all, thank you for replying to my pr. After reading your reply and reviewing the relevant code, I think `the current working directory` is already very appropriate, but I think it may be easier for readers to understand it if it is changed to `the current working directory of the Node.js process`. If you still think the former is better, then I will change to the former. Thank you again for replying to my pr, you are the first person to reply to my pr.