docs(configuration): Update docs for entry-context (#5778)
* Update docs for entry-context

I think this current directory has a ambiguity. Does the current directory refer to the root directory of the project? Or is it the directory where webpack.config.js is located?

After I tested it, I found neither of them, the default should be the directory of the script that starts webpack. For example, I ran the script `webpack --config relativebase` in a non-project root directory. Then, the value of the context at this time is the directory where the script is located.

In order to better illustrate my ideas, I wrote a demo code, but I don’t know how to give the sample code, so I created a warehouse to store the sample code. https://github.com/spicylemonhaha/webpack-demo

* Update docs for entry-context

First of all, thank you for replying to my pr. After reading your reply and reviewing the relevant code, I think `the current working directory` is already very appropriate, but I think it may be easier for readers to understand it if it is changed to `the current working directory of the Node.js process`. If you still think the former is better, then I will change to the former. Thank you again for replying to my pr, you are the first person to reply to my pr.

* Update docs for entry-context

Undo some of the previous changes

* Update docs for entry-context

Text alignment

* Update docs for entry-context

remove process

* Update src/content/configuration/entry-context.mdx

Co-authored-by: Sam Chen <chenxsan@gmail.com>