docs(tree-shaking): remove imprecise language
The prior language suggested that ESM helps identify side effects. This is not the case.

The following case demonstrates.

```js
export const foo = (() => { window.bar = true; return 0; })();
```

Module evaluation of the above ES module induces a side effect during evaluation, and usage of ESM did not help identify it.

ESM did help compiler tooling evaluate `usedExports`, because the exported symbols are fixed.

The use of PURE and sideEffects are a bit contradictory in the webpack definitions. It's probably worth defining explicit definitions for webpack's interpreation of `pure` and `side effect` in the glossary page.