Clarify interaction between `--node-env` and `mode`
The interpretation of the current description is ambiguous, as it pertains to how a configuration function can or cannot benefit from the use of `--node-env` as a way to set `mode`. One might interpret it as if that happens before the configuration function is invoked, and therefore should be available to use for logic like switching in the function. That would be wrong. This update makes it clear that the assignment of a default `mode` based on `--node-env` happens after the configuration function is executed.

Related: #4009 

Fixes #3992