docs: add note about how publicPath interacts with asset modules
This updates the documentation to better define how `publicPath` interacts with `asset` module types. Specifically, it is poorly documented on how to dynamically set the `publicPath` on the fly for these module types. This is important and useful in an enterprise environment where there may be different upstream CDNs or other locations that the app hosts assets on that aren't the server or a static site.